requirements.txt
einops==0.8.2
fvcore==0.1.5.post20221221
joblib==1.5.3
numpy==2.4.2
opencv_python==4.13.0.90
pandas==3.0.0
Pillow==12.1.1
pytorchvideo==0.1.5
rich==14.3.2
scikit_learn==1.8.0
tensorboardx==2.6.4
timm==1.0.24
torch==2.10.0
torchfitter==4.3.0
torchvision==0.25.0
tqdm==4.67.3
